Update Radio Group and Icon
